Initial CheckIn
This is a basic setup of Backend Serivce project in spring data rest.

This project can be downloaded and open it in Eclipse or STS and build
and run from STS Spring boot server or Maven(goto repo root dir where
you can find the pom.xml in cmd >run command <mvn clean package> this
will create .jar in the target dir of project).

Download the project repo to local and To Import to Eclipse from Eclipse
> New > Import > Existing Maven project > browse to the root of project
dir > select the pom.xml